working version

+++ router/user.py
@@ -6,21 +6,25 @@ import schemas
 from database import get_async_session
 from users import current_active_user
 
+#user_router является объектом маршрутизатора
 user_router = APIRouter(prefix='/user')
 
-
+#endpoint
 @user_router.get("/all", response_model=list[schemas.UserRead])
 async def get_all_users(db=Depends(get_async_session)):
     try:
-        db_all_users = await crud.get_all_users(db)
-        return db_all_users
+        db_all_users = await crud.get_all_users(db) # Вызов функции get_all_users из модуля crud для получения всех пользователей из базы данных
+        return db_all_users # Возвращение списка пользователей в качестве ответа
     except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
+        raise HTTPException(status_code=500, detail=str(e)) # Если произошла ошибка, генерирование исключения HTTPException с кодом состояния 500 и сообщением об ошибке
 
-
+#endpoint
 @user_router.get("/photo/get")
-async def get_user_photo(user_id: str):
-    file_path = f"photos/user_{user_id}.png"
+async def get_user_photo(user_id: str, db = Depends(get_async_session)):
+    if not await crud.is_user_exist(db, user_id):
+        raise HTTPException(status_code=403, detail="User not found")
+
+    file_path = f"photo/user_{user_id}.png"
     try:
         with open(file_path, "rb") as file:
             photo_data = file.read()
@@ -30,10 +34,13 @@ async def get_user_photo(user_id: str):
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-
+#endpoint
 @user_router.post("/photo/set")
 async def set_user_photo(user=Depends(current_active_user), photo: UploadFile = File(...)):
-    file_path = f"photos/user_{user.id}.png"
+    if not await crud.is_user_exist(db, user.id):
+        raise HTTPException(status_code=403, detail="User not found")
+
+    file_path = f"photo/user_{user.id}.png"
     try:
         with open(file_path, "wb") as file:
             file.write(await photo.read())
@@ -41,9 +48,12 @@ async def set_user_photo(user=Depends(current_active_user), photo: UploadFile = 
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-
+#endpoint
 @user_router.get("/info/get", response_model=schemas.GetUserInfo)
 async def get_user_info(user_id: str, db=Depends(get_async_session)):
+    if not await crud.is_user_exist(db, user_id):
+        raise HTTPException(status_code=403, detail="User not found")
+
     try:
         db_user_info = await crud.get_user_info(db, user_id)
 
@@ -56,21 +66,30 @@ async def get_user_info(user_id: str, db=Depends(get_async_session)):
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-
+#endpoint
 @user_router.post("/info/set")
 async def set_user_info(user_info: schemas.CreateUserInfo, user=Depends(current_active_user),
                         db=Depends(get_async_session)):
+    if not await crud.is_user_exist(db, user.id):
+        raise HTTPException(status_code=403, detail="User not found")
+
+    if await crud.is_user_info_exist(db, user.id):
+        raise HTTPException(status_code=403, detail="User info already exist")
+
     try:
         await crud.create_user_info(db, user.id, user_info)
         return "User info successfully setted"
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-
+#endpoint
 @user_router.patch("/info/update")
 async def update_user_info(user_info: schemas.UpdateUserInfo,
                            user=Depends(current_active_user),
                            db=Depends(get_async_session),):
+    if not await crud.is_user_exist(db, user.id):
+        raise HTTPException(status_code=403, detail="User not found")
+
     try:
         await crud.update_user_info(db, user.id, user_info)
         return "User info successfully updated"
